requirements.txt
certifi==2018.4.16
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
RPi.GPIO==0.6.5
six==1.12.0
Werkzeug==0.15.2
